chg: replace debug with winston for logging in all classes
