+++ src/utilities/socket-server.js
@@ -1,10 +1,8 @@
 'use strict'
 
 const EventEmitter = require('events')
-const debug = require( 'debug')
 const WebSocketServer = require('ws').Server
-
-const log = debug('oracle:socketserver')
+const logger = require('../logger.js');
 
 module.exports = class SocketServer extends EventEmitter {
   constructor() {
@@ -23,16 +21,16 @@ module.exports = class SocketServer extends EventEmitter {
                   pubsub.publish(ws, json.channel, json.message)
                   break
                 case 'SUBSCRIBE':
-                  log('SUBSCRIBE: ' + json.message)
+                  logger.debug('SUBSCRIBE: ' + json.message)
                   pubsub.subscribe(ws, json.channel)
                   
                   // if (process.env.ALLOW_CORS) {
-                  //   log('client subscribed to: ' + json.channel)
+                  //   logger.debug('client subscribed to: ' + json.channel)
                   //   pubsub.subscribe(ws, json.channel)
                   // }
                   // else {
                   //   if (req.headers.origin != process.env.DOMAIN) {
-                  //     log('header denied: ' + req.headers.origin + ' :' + json.message)
+                  //     logger.debug('header denied: ' + req.headers.origin + ' :' + json.message)
                   //     ws.send(JSON.stringify({ access: 'denied' }))
                   //     ws.close()
                   //   }
@@ -40,15 +38,15 @@ module.exports = class SocketServer extends EventEmitter {
                   break
               }
             } catch (error) {
-              log(error)
+              logger.error(error)
             }
           })
           ws.on('close', () => {
-            log('Stopping client connection.')
+            logger.debug('Stopping client connection.')
           })
           ws.on('error', (error) => {
-            log('SocketServer error')
-            log(error)
+            logger.error('SocketServer error')
+            logger.error(error)
           })
         })
       }
